Swift settings, clean up storage protocols
