+++ Package.swift
@@ -1,8 +1,19 @@
-// swift-tools-version: 6.0
+// swift-tools-version: 6.1
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import Foundation
 import PackageDescription
+
+var defaultSwiftSettings: [SwiftSetting] = [
+    // https://github.com/swiftlang/swift-evolution/blob/main/proposals/0335-existential-any.md
+    .enableUpcomingFeature("ExistentialAny"),
+
+    // https://github.com/swiftlang/swift-evolution/blob/main/proposals/0409-access-level-on-imports.md
+    .enableUpcomingFeature("InternalImportsByDefault"),
+
+    // https://github.com/swiftlang/swift-evolution/blob/main/proposals/0444-member-import-visibility.md
+    .enableUpcomingFeature("MemberImportVisibility"),
+]
 
 let package = Package(
     name: "swift-zip-archive",
@@ -20,7 +31,8 @@ let package = Package(
             dependencies: [
                 "CZipArchiveZlib",
                 .product(name: "SystemPackage", package: "swift-system"),
-            ]
+            ],
+            swiftSettings: defaultSwiftSettings
         ),
         .testTarget(
             name: "ZipArchiveTests",
